% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safesite.R
\name{safesite}
\alias{safesite}
\title{Custom site generator to safely build R Markdown website}
\usage{
safesite(input, encoding = getOption("encoding"), ...)
}
\description{
The site generator \code{safesite} modifies the render function produced by
the site generator \code{rmarkdown:::default_site} (see
\code{\link[rmarkdown]{render_site}}) to run \code{\link[rmarkdown]{render}}
in a separate R process using the R package \link{callr}.
}
\details{
To use this alternative site generator, add the following line to the YAML
header of \code{index.Rmd}:

\preformatted{site: safesite::safesite}
}
